fetch_api/requirements.txt
django-filter==2.4.0
djangorestframework==3.12.1
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
